requirements.txt
Django==1.11.3
django-crispy-forms==1.6.1
django-simple-history==1.9.0
psycopg2==2.7.3
pytz==2017.2
django-rest-framework==3.7.3